backend/app/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.3